Finished game.

+++ src/edu/purdue/cs/dodgeball/Player.java
@@ -1,27 +1,27 @@
 package edu.purdue.cs.dodgeball;
 
 public class Player {
-	private int x;
-	private int y;
+	private float x;
+	private float y;
 	
-	public Player(int x, int y) {
+	public Player(float x, float y) {
 		this.x = x;
 		this.y = y;
 	}
 	
-	public int getX() {
+	public float getX() {
 		return x;
 	}
 	
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 	
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 	
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y = y;
 	}
 }
